counter helper placed within counter
