+++ src/model/mod.rs
@@ -1,7 +1,5 @@
 /// Arbitrage chance
 pub mod chance;
-/// System MPS counter
-pub mod counter;
 /// Order
 pub mod order;
 /// Orderbook
